pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "x-vc"
version = "0.1.0"
description = "Zero-shot Streaming Voice Conversion in Codec Space"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "einops>=0.7.0",
    "x-transformers>=1.40.0",
    "hydra-core>=1.3.0",
    "julius>=0.2.7",
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "omegaconf>=2.3.0",
    "scipy>=1.10.0",
    "soundfile>=0.12.0",
    "soxr>=0.3.0",
    "tqdm>=4.60.0",
    "einx>=0.3.0",
    "transformers>=4.44.0",
    "packaging>=21.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "descript-audiotools",
    "modelscope>=1.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["x_vc*"]
exclude = [".venv*"]
